test(bedrock): split cost utility tests
