+++ specs/GH727/tasks.md
@@ -42,23 +42,27 @@ GH-727 / #727
 - [x] `SP727-T29` Owner: coordinator. Done when: `src/core/providers/azure/responses/processor.rs` delegates tests with `#[path = "processor_tests.rs"] mod tests;`. Verify: `git diff -- src/core/providers/azure/responses/processor.rs`.
 - [x] `SP727-T30` Owner: coordinator. Done when: original inline Azure responses processor tests are moved to `src/core/providers/azure/responses/processor_tests.rs` without changing assertions. Verify: `rg -n "fn test_" src/core/providers/azure/responses/processor_tests.rs`.
 - [x] `SP727-T31` Owner: verification owner. Done when: both Azure responses processor files are below U-16's 800-line ceiling and focused tests pass. Verify: `wc -l src/core/providers/azure/responses/processor.rs src/core/providers/azure/responses/processor_tests.rs`; `cargo test core::providers::azure::responses::processor --lib --all-features`.
-- [ ] `SP727-T32` Owner: verification owner. Done when: SpecRail, formatting, all-features check, PR CI, and review-thread gate pass for the Azure responses processor tranche. Verify: `SPEC_RAIL=/path/to/specrail; python3 "$SPEC_RAIL/checks/check_workflow.py" --repo "$SPEC_RAIL" --spec-dir "$PWD/specs/GH727"`; `cargo fmt --all -- --check`; `cargo check --all-features --locked`; GitHub PR CI and review-thread query.
+- [x] `SP727-T32` Owner: verification owner. Done when: SpecRail, formatting, all-features check, PR CI, and review-thread gate pass for the Azure responses processor tranche. Verify: #810 PR body, green PR CI, and GraphQL review-thread query.
+- [x] `SP727-T33` Owner: coordinator. Done when: `src/core/providers/bedrock/utils/cost.rs` delegates tests with `#[path = "cost_tests.rs"] mod tests;`. Verify: `git diff -- src/core/providers/bedrock/utils/cost.rs`.
+- [x] `SP727-T34` Owner: coordinator. Done when: original inline Bedrock cost tests are moved to `src/core/providers/bedrock/utils/cost_tests.rs` without changing assertions. Verify: `rg -n "fn test_" src/core/providers/bedrock/utils/cost_tests.rs`.
+- [x] `SP727-T35` Owner: verification owner. Done when: both Bedrock cost files are below U-16's 800-line ceiling and focused tests pass. Verify: `wc -l src/core/providers/bedrock/utils/cost.rs src/core/providers/bedrock/utils/cost_tests.rs`; `cargo test core::providers::bedrock::utils::cost --lib --all-features`.
+- [ ] `SP727-T36` Owner: verification owner. Done when: SpecRail, formatting, all-features check, PR CI, and review-thread gate pass for the Bedrock cost tranche. Verify: `SPEC_RAIL=/path/to/specrail; python3 "$SPEC_RAIL/checks/check_workflow.py" --repo "$SPEC_RAIL" --spec-dir "$PWD/specs/GH727"`; `cargo fmt --all -- --check`; `cargo check --all-features --locked`; GitHub PR CI and review-thread query.
 
 ## 并行拆分
 
-This is a serial writable lane for one Azure responses processor test file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
+This is a serial writable lane for one Bedrock cost test file family. Other #727 large-file tranches may be planned read-only in parallel, but they must not edit this branch.
 
 Writable ownership for this lane:
 
 - `specs/GH727/`
-- `src/core/providers/azure/responses/processor.rs`
-- `src/core/providers/azure/responses/processor_tests.rs`
+- `src/core/providers/bedrock/utils/cost.rs`
+- `src/core/providers/bedrock/utils/cost_tests.rs`
 
 ## 验证
 
 - SpecRail packet validation.
 - `cargo fmt --all -- --check`
-- `cargo test core::providers::azure::responses::processor --lib --all-features`
+- `cargo test core::providers::bedrock::utils::cost --lib --all-features`
 - `cargo check --all-features --locked`
 - PR CI and GraphQL review-thread gate before merge.
